### 2021-09-22 #3755 @whedon
[PRE REVIEW]: QuasinormalModes.jl: A Julia package for computing1 discrete eigenvalues of second order ODEs
**Submitting author:** @lucass-carneiro (<a href="http://orcid.org/0000-0001-6764-1812">Lucas Timotheo Sanches</a>)
**Repository:** <a href="https://github.com/lucass-carneiro/QuasinormalModes.jl/" target ="_blank">https://github.com/lucass-carneiro/QuasinormalModes.jl/</a>
**Version:** v1.0.0
