### 2021-09-29 #3755 @ygrange
It's a bit outside my field (I'm more of an experimental astronomy guy) but I'd be happy to poke at it if someone with more (theoretical astronomy/physics) domain knowledge could supplement that.
