### 2021-12-06 #3755 @pdebuyl
Hi @ziotom78 would you review the article "QuasinormalModes.jl: A Julia package for computing1 discrete eigenvalues of second order ODEs" for the Journal of Open Source Software?
