### 2021-12-06 #3755 @ziotom78
Hi @pdebuyl, as I wrote above, my domain is instrumental astronomy, so theoretical models of Schwartzschild black holes is a bit outside my domain. I can co-referee this paper if I am sided by somebody with a more theoretical background.
