### 2022-01-20 #4077 @whedon
[REVIEW]: QuasinormalModes.jl: A Julia package for computing1 discrete eigenvalues of second order ODEs
**Submitting author:** @lucass-carneiro (<a href="http://orcid.org/0000-0001-6764-1812">Lucas Timotheo Sanches</a>)
**Repository:** https://github.com/lucass-carneiro/QuasinormalModes.jl/
**Branch with paper.md** (empty if default branch): master
**Version:** v1.1.0
