### 2022-02-04 #4077 @cescalara
Hi @lucass-carneiro,

Thanks for your work on `QuasinormalModes.jl`. I started the review today, here are my questions/comments:

- At the moment it is not clear which dependencies are needed to run the tests/docs asopposed to the core code. The latest recommendation is to use separate `Project.toml` files inside `tests/` and `docs` to handle this - see https://pkgdocs.julialang.org/v1/creating-packages/#Adding-tests-to-the-package for more info.

- There are logging statements available in Julia such as `@info` and `@warn` that have a lot of advantages over a call to `println()` - see https://docs.julialang.org/en/v1/stdlib/Logging/. These could be used when not using throwing exceptions, e.g. [here]( https://github.com/lucass-carneiro/QuasinormalModes.jl/blob/f22d12304b247415df94483fc4f7355fb8fa0393/src/eigenvalue_computing/compute_eigenvalues.jl#L37).

- I noticed that the docs and example scripts are out of sync, e.g. I need to add some calls to `BigFloat()` in when constructing problems in the Schwarzschild example to avoid an error. A nice way to avoid manually updating both is to use [Literate.jl](https://fredrikekre.github.io/Literate.jl/v2/) to convert your scripts to doc notebooks that are actually run when the docs are built, so outputs will also be displayed for users to compare with.

- Can you provide the code used to make the benchmark plots shown in the paper? This would make it easier for me to check the performance claims. It could also be useful to have it in the repo, as an example or test.

- Could you add some guidelines on a good workflow for ensuring enough iterations in the AIM algorithm when applying this method generally? Is it enough just to check the final output from `computeEigenvalues()`?

- In the paper, some in-text citations have parentheses around them. I think you can just remove the square brackets to fix this.

- More superficial suggestions:
  - It would be nice to have "oscillator" spelt correctly throughout the code/docs.
  - Also FYI, the Julia style guide for function naming is lowercase with underscores (https://docs.julialang.org/en/v1/manual/style-guide/#Use-naming-conventions-consistent-with-Julia-base/), but I don't think this is necessary for JOSS.

I hope this is useful and am happy to discuss more.
