### 2022-02-07 6ff41a9
Added a new folder with files for generating and processing benchmark results presented in the JOSS paper
