### 2022-02-07 6090ff7
Added a separate Project.toml and Manifest.toml file for the test suite as recommended by Julia's official documentation
