### 2022-02-07 a80ca90
Updated the documentation source files for the usage examples to sync them to the actual source files contained in the examples/ folder
