+++ docs/src/schw.md
@@ -34,7 +34,14 @@ which implies that when ``r=1`` we have ``x=0`` and when ``r\rightarrow\infty`` 
 
 # Implementing the master equation as an analytic problem
 
-In this section we will assume that the user has installed and loaded the `SymEngine` package. The first step is to create a parametric type that sub-types `AnalyticAIMProblem`. As the eigenvalue in the master equation is a quadratic polynomial, we will sub-type `QuadraticEigenvalueProblem` with the following structure:
+The first step is to load the required packages to run this example: `QuasinormalModes` and `SymEngine`:
+
+```julia
+using QuasinormalModes
+using SymEngine
+```
+
+Next, we create a parametric type that sub-types `AnalyticAIMProblem`. As the eigenvalue in the master equation is a quadratic polynomial, we will sub-type `QuadraticEigenvalueProblem` with the following structure:
 
 ```julia
 struct SchwarzschildData{N,T} <: QuadraticEigenvalueProblem{N,T}
@@ -61,7 +68,7 @@ function SchwarzschildData(nIter::N, x0::T, l::N, s::N) where {N,T}
 end
 ```
 
-This constructor can be used by passing the values directly instead of explicitly declaring type parameters. The final step is to extend the default accessors functions to operate on `SchwarzschildData`
+This constructor can be used by passing the values directly instead of explicitly declaring type parameters. The final step is to extend the default accessors functions to operate on `SchwarzschildData`:
 
 ```julia
 QuasinormalModes.λ0(d::SchwarzschildData{N,T}) where {N,T} = d.exprs[1]
@@ -73,7 +80,7 @@ QuasinormalModes.get_x0(d::SchwarzschildData{N,T}) where {N,T} = d.x0
 QuasinormalModes.get_ODEvar(d::SchwarzschildData{N,T}) where {N,T} = d.vars[1]
 QuasinormalModes.get_ODEeigen(d::SchwarzschildData{N,T}) where {N,T} = d.vars[2]
 ```
-These functions are fairly straightforward and accessors and require no additional comment.
+These functions are fairly straightforward accessors and require no additional comment.
 
 # Implementing the master equation as a numeric problem
 
@@ -90,7 +97,7 @@ end
 
 Here `nIter` and `x0` have the same meaning as before, but now instead of storing symbolic variables and expressions we store two additional unsigned integers, `l` and `s`. These are the angular and spin parameters of the master equation. Here we must store them in the struct as they can't be "embedded" into the expressions for `λ0` and `S0` as in the analytic case.
 
-We proceed once again by creating a more convenient constructor. This time no intermediate computation is required upon the construction:
+We proceed once again by creating a more convenient constructor. This time no intermediate computation is required upon construction:
 
 ```julia
 function NSchwarzschildData(nIter::N, x0::T, l::N, s::N) where {N,T}
@@ -115,7 +122,7 @@ This time `λ0` and `S0` return two parameters lambda functions that will be cal
 We create our problems and cache objects by calling the constructors:
 
 ```julia
-p_ana = SchwarzschildData(0x00030, Complex(0.43, 0.0), 0x00000, 0x00000);
+p_ana = SchwarzschildData(0x00030, Complex(BigFloat("0.43"), BigFloat("0.0")), 0x00000, 0x00000);
 p_num = NSchwarzschildData(0x00030, Complex(0.43, 0.0), 0x00000, 0x00000);
 
 c_ana = AIMCache(p_ana)
@@ -131,7 +138,7 @@ To compute eigenvalues, 3 functions are provided:
 2. `computeEigenvalues`: Compute a single, or a list of eigenvalues.
 3. `eigenvaluesInGrid`: Find all eigenvalues in a certain numerical grid.
 
[…]
