### 2022-02-07 9a28313
Fixed the typo across the whole codbase where oscillator was written with only one l
