+++ test/runtests.jl
@@ -1,6 +1,10 @@
 using SafeTestsets
 
 @time begin
-    @time @safetestset "Quantum Harmonic oscilator tests" begin include("harmonic_oscilator.jl") end
-    @time @safetestset "Schwarzschild quasinormal modes tests" begin include("schwarzschild_qnm.jl") end
+    @time @safetestset "Quantum Harmonic oscillator tests" begin
+        include("harmonic_oscillator.jl")
+    end
+    @time @safetestset "Schwarzschild quasinormal modes tests" begin
+        include("schwarzschild_qnm.jl")
+    end
 end
