+++ examples/harmonic_oscillator.jl
@@ -2,29 +2,29 @@ using QuasinormalModes
 using SymEngine
 
 # ------------------------------------------------------------------
-# 1. Analytic harmonic oscilator
+# 1. Analytic harmonic oscillator
 # ------------------------------------------------------------------
 
 struct HarmonicOscilatorData{N,T} <: QuadraticEigenvalueProblem{N,T}
     nIter::N
     x0::T
 
-    vars::Tuple{Basic, Basic}
-    exprs::Tuple{Basic, Basic}
+    vars::Tuple{Basic,Basic}
+    exprs::Tuple{Basic,Basic}
 end
 
 function HarmonicOscilatorData(nIter::N, x0::T) where {N,T}
-	
+
     vars = @vars x ω
 
-    λ0 = 2*x
+    λ0 = 2 * x
     S0 = 1 - ω
 
     return HarmonicOscilatorData{N,T}(nIter, x0, vars, (λ0, S0))
 end
 
 QuasinormalModes.λ0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.exprs[1]
-QuasinormalModes.S0(d::HarmonicOscilatorData{N,T}) where {N,T}  = d.exprs[2]
+QuasinormalModes.S0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.exprs[2]
 
 QuasinormalModes.get_niter(d::HarmonicOscilatorData{N,T}) where {N,T} = d.nIter
 QuasinormalModes.get_x0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.x0
@@ -33,7 +33,7 @@ QuasinormalModes.get_ODEvar(d::HarmonicOscilatorData{N,T}) where {N,T} = d.vars[
 QuasinormalModes.get_ODEeigen(d::HarmonicOscilatorData{N,T}) where {N,T} = d.vars[2]
 
 # ------------------------------------------------------------------
-# 2. Numeric harmonic oscilator
+# 2. Numeric harmonic oscillator
 # ------------------------------------------------------------------
 
 struct NHarmonicOscilatorData{N,T} <: NumericAIMProblem{N,T}
@@ -45,8 +45,8 @@ function NHarmonicOscilatorData(nIter::N, x0::T) where {N,T}
     return NHarmonicOscilatorData{N,T}(nIter, x0)
 end
 
-QuasinormalModes.λ0(::NHarmonicOscilatorData{N,T}) where {N,T} = (x,ω) -> 2*x
-QuasinormalModes.S0(::NHarmonicOscilatorData{N,T}) where {N,T} = (x,ω) -> 1 - ω + x - x
+QuasinormalModes.λ0(::NHarmonicOscilatorData{N,T}) where {N,T} = (x, ω) -> 2 * x
+QuasinormalModes.S0(::NHarmonicOscilatorData{N,T}) where {N,T} = (x, ω) -> 1 - ω + x - x
 
 QuasinormalModes.get_niter(d::NHarmonicOscilatorData{N,T}) where {N,T} = d.nIter
 QuasinormalModes.get_x0(d::NHarmonicOscilatorData{N,T}) where {N,T} = d.x0
@@ -74,7 +74,7 @@ function printEigen(eigenvalues)
     for i in eachindex(eigenvalues)
         println("n = $i, ω = $(eigenvalues[i])")
     end
-    
+
     println("--------------------------------------")
 
     return nothing
