+++ test/harmonic_oscillator.jl
@@ -3,7 +3,7 @@ using SymEngine
 using Test
 
 # ------------------------------------------------------------------
-# 1. Analytic harmonic oscilator
+# 1. Analytic harmonic oscillator
 # ------------------------------------------------------------------
 
 struct HarmonicOscilatorData{N,T} <: QuadraticEigenvalueProblem{N,T}
@@ -15,7 +15,7 @@ struct HarmonicOscilatorData{N,T} <: QuadraticEigenvalueProblem{N,T}
 end
 
 function HarmonicOscilatorData(nIter::N, x0::T) where {N,T}
-	
+
     vars = @vars x ω
 
     λ0 = 2 * x
@@ -25,7 +25,7 @@ function HarmonicOscilatorData(nIter::N, x0::T) where {N,T}
 end
 
 QuasinormalModes.λ0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.exprs[1]
-QuasinormalModes.S0(d::HarmonicOscilatorData{N,T}) where {N,T}  = d.exprs[2]
+QuasinormalModes.S0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.exprs[2]
 
 QuasinormalModes.get_niter(d::HarmonicOscilatorData{N,T}) where {N,T} = d.nIter
 QuasinormalModes.get_x0(d::HarmonicOscilatorData{N,T}) where {N,T} = d.x0
@@ -34,7 +34,7 @@ QuasinormalModes.get_ODEvar(d::HarmonicOscilatorData{N,T}) where {N,T} = d.vars[
 QuasinormalModes.get_ODEeigen(d::HarmonicOscilatorData{N,T}) where {N,T} = d.vars[2]
 
 # ------------------------------------------------------------------
-# 2. Numeric harmonic oscilator
+# 2. Numeric harmonic oscillator
 # ------------------------------------------------------------------
 
 struct NHarmonicOscilatorData{N,T} <: NumericAIMProblem{N,T}
@@ -57,11 +57,11 @@ QuasinormalModes.get_x0(d::NHarmonicOscilatorData{N,T}) where {N,T} = d.x0
 # ------------------------------------------------------------------
 
 @testset "Serial Numeric correctness - 10 iterations" begin
-    p = NHarmonicOscilatorData(0x0000A, 0.5);
+    p = NHarmonicOscilatorData(0x0000A, 0.5)
     c = AIMCache(p)
     ev = eigenvaluesInGrid(Serial(), p, c, (0.0, 20.0))
     ce = [computeEigenvalues(Serial(), p, c, guess) for guess in 1.0:2.0:19.0]
-    
+
     @test length(ev) == 10
     @test round(ev[1]) == 1
     @test round(ev[2]) == 3
@@ -87,12 +87,12 @@ QuasinormalModes.get_x0(d::NHarmonicOscilatorData{N,T}) where {N,T} = d.x0
     @test round(ce[10]) == 19
 end
 
- @testset "Serial Numeric correctness - 20 iterations" begin
-    p = NHarmonicOscilatorData(0x00014, 0.5);
+@testset "Serial Numeric correctness - 20 iterations" begin
+    p = NHarmonicOscilatorData(0x00014, 0.5)
     c = AIMCache(p)
     ev = eigenvaluesInGrid(Serial(), p, c, (0.0, 20.0))
     ce = [computeEigenvalues(Serial(), p, c, guess) for guess in 1.0:2.0:19.0]
-    
+
     @test length(ev) == 10
     @test round(ev[1]) == 1
     @test round(ev[2]) == 3
@@ -119,11 +119,11 @@ end
 end
 
 @testset "Serial Numeric correctness - 50 iterations" begin
-    p = NHarmonicOscilatorData(0x00032, 0.5);
[…]
