### 2022-02-07 3bdbd16
Added an independent Project.toml and Manifest.toml to the documentation
