### 2022-02-07 ade4f8d
Removed dead code from the documentatiob makefile
