+++ docs/make.jl
@@ -1,13 +1,6 @@
 # To view in browser start a server in the build dir:
 # python -m http.server --bind localhost
 
-#if !("./" in LOAD_PATH)
-#    push!(LOAD_PATH, "./")
-#end
-
-#using Pkg
-#Pkg.add("Documenter")
-
 using Documenter
 using QuasinormalModes
 
@@ -22,10 +15,3 @@ makedocs(sitename = "QuasinormalModes.jl",
         "api_ref.md"
     ]
 )
-
-# deploydocs(
-#     repo = "github.com/lucass-carneiro/QuasinormalModes.jl.git",
-#     target = "build",
-#     devurl = "./",
-#     push_preview = false,
-# )
