### 2022-02-07 09cb3f0
Updated the gh-pages repository to reflect the updated exemple section in the documentation
