+++ search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"sho/#Complete-Example:-The-Harmonic-Oscilator","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"We will now turn away from general relativity and use QuasinormalModes.jl to compute to compute the energy eigenvalues of the quantum harmonic oscillator following this paper.","category":"page"},{"location":"sho/#Mathematical-preliminaries","page":"Complete Example: The Harmonic Oscilator","title":"Mathematical preliminaries","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"If we measure the energy of the system in units of hbaromega and distance in units of sqrthbar(momega) the time independent Schrödinger equation for the quantum harmonic oscilator is written as","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"-psi^primeprime(x) + x^2psi(x) = epsilonpsi(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"where we defined epsilon equiv 2 E and E is the quantum state's energy. Imposing that psi(x) decays like a Gaussian distribution asymptotically, we apply the ansatz","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"psi(x) = e^-x^22f(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"which, substituting in the original equation, yields","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"f^primeprime(x) = 2 x f^prime(x) + (1-epsilon)f(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"This allows us to easily identify lambda_0 = 2x and s_0 = 1 - epsilon. In all our implementations we shall refer the sought eigenvalue epsilon using the variable ω in order to maintain consistency with the previous example.","category":"page"},{"location":"sho/#Implementing-the-master-equation-as-an-analytic-problem","page":"Complete Example: The Harmonic Oscilator","title":"Implementing the master equation as an analytic problem","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"In this section we will assume that the user has installed and loaded the SymEngine package. The first step is to create a parametric type that sub-types AnalyticAIMProblem. As the eigenvalue in the master equation is a quadratic polynomial, we will sub-type QuadraticEigenvalueProblem with the following structure:","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","text":"struct HarmonicOscilatorData{N,T} <: QuadraticEigenvalueProblem{N,T}\n    nIter::N\n    x0::T\n\n    vars::Tuple{Basic, Basic}\n    exprs::Tuple{Basic, Basic}\nend","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscilator","title":"Complete Example: The Harmonic Oscilator","te
[…]
