### 2022-02-08 7510b72
Added a new documentation section providing guidelines on a good workflow for working with the package
