+++ docs/src/index.md
@@ -1,5 +1,5 @@
 # Table of Contents
 
 ```@contents
-Pages = ["intro.md", "schw.md", "sho.md", "api_ref.md"]
+Pages = ["intro.md", "org.md", "schw.md", "sho.md", "api_ref.md"]
 ```
