+++ docs/src/org.md
@@ -65,7 +65,7 @@ For problems with the `IsAnalytic` trait, the user must implement the following 
 1. `get_ODEvar` which returns an object that represents the ODE's variable.
 2. `get_ODEeigen` which returns an object that represents the ODE's eigenvalue.
 
-Failure to implement these functions returns an error with the appropriate message. Note that these traits only check that such functions are implemented for a certain problem type and not that they follow a particular implementation pattern. The contract on the functions implementations is *soft* and will be clarified further on. Failure to abide by these soft contracts results in undefined behaviour.
+Failure to implement these functions returns an error with the appropriate message. Note that these traits only check that such functions are implemented for a certain problem type and not that they follow a particular implementation pattern. The contract on the functions implementations is *soft* and will be clarified further on. Failure to abide by these soft contracts results in undefined behavior.
 
 ## Extending the default functionality
 
@@ -94,7 +94,33 @@ In order to minimize memory allocations, all functions that actually compute eig
 
 ## Stepping methods
 
-In order to compute ``\delta_n``, `QuasinormalModes.jl` evolves ``\lambda_0`` and ``s_0`` according to the previously stated equations. The evolution from step 1 to step ``n`` must happen sequentially but the step itself, that is, the computation of new values of ``\lambda`` and ``s`` from old ones can be performed in parallel. We've provided singleton types that allow the user to control this behaviour by passing instances of those types to the eigenvalue computing functions. The user can currently choose the following stepping methods:
+In order to compute ``\delta_n``, `QuasinormalModes.jl` evolves ``\lambda_0`` and ``s_0`` according to the previously stated equations. The evolution from step 1 to step ``n`` must happen sequentially but the step itself, that is, the computation of new values of ``\lambda`` and ``s`` from old ones can be performed in parallel. We've provided singleton types that allow the user to control this behavior by passing instances of those types to the eigenvalue computing functions. The user can currently choose the following stepping methods:
 
 1. `Serial`: Each instruction in a single AIM step is executed sequentially.
 2. `Threaded`: Instruction in a single AIM step is executed in parallel using Julia's built-in `Threads.@threads` macro.
+
+## Computing eigenvalues and general workflow guidelines 
+
+To compute eigenvalues, 3 functions are provided:
+1. `computeDelta!`: Computes the AIM "quantization condition" ``\delta_n``.
+2. `computeEigenvalues`: Computes a single, or a list of eigenvalues.
+3. `eigenvaluesInGrid`: Find all eigenvalues in a certain numerical grid.
+
+Depending on the problem type, these functions return and behave differently. In a `QuadraticEigenvalueProblem`.
+1. `computeDelta!`: Returns a polynomial whose roots are the eigenvalues of the ODE.
+2. `computeEigenvalues`: Computes the complete list of eigenvalues given by the roots of the computed polynomial.
+
+In a `NumericAIMProblem`,
+1. `computeDelta!`: Returns a value of the quantization condition at a given point in the complex plane.
+2. `computeEigenvalues`: Computes a single eigenvalues from an initial trial frequency.
+3. `eigenvaluesInGrid`: Attempts to find eigenvalues using a grid of real or complex data points as initial trial frequencies passed to `NLSolve`.
+
+For more detail on these functions and their behaviors with each problem type refer to the the [API Reference](api_ref.md) where specific descriptions can be found. 
+
[…]
