### 2022-02-08 f98a981
Added a new documentation section providing guidelines on a good workflow for working with the package
