+++ versions.js
@@ -1,6 +1,0 @@
-var DOC_VERSIONS = [
-  "stable",
-  "v1.0",
-];
-var DOCUMENTER_NEWEST = "v1.0.0";
-var DOCUMENTER_STABLE = "stable";
