+++ search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"sho/#Complete-Example:-The-Harmonic-Oscillator","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"We will now turn away from general relativity and use QuasinormalModes.jl to compute to compute the energy eigenvalues of the quantum harmonic oscillator following this paper.","category":"page"},{"location":"sho/#Mathematical-preliminaries","page":"Complete Example: The Harmonic Oscillator","title":"Mathematical preliminaries","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"If we measure the energy of the system in units of hbaromega and distance in units of sqrthbar(momega) the time independent Schrödinger equation for the quantum harmonic oscillator is written as","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"-psi^primeprime(x) + x^2psi(x) = epsilonpsi(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"where we defined epsilon equiv 2 E and E is the quantum state's energy. Imposing that psi(x) decays like a Gaussian distribution asymptotically, we apply the ansatz","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"psi(x) = e^-x^22f(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"which, substituting in the original equation, yields","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"f^primeprime(x) = 2 x f^prime(x) + (1-epsilon)f(x)","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"This allows us to easily identify lambda_0 = 2x and s_0 = 1 - epsilon. In all our implementations we shall refer the sought eigenvalue epsilon using the variable ω in order to maintain consistency with the previous example.","category":"page"},{"location":"sho/#Implementing-the-master-equation-as-an-analytic-problem","page":"Complete Example: The Harmonic Oscillator","title":"Implementing the master equation as an analytic problem","text":"","category":"section"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"The first step is to load the required packages to run this example: QuasinormalModes and SymEngine:","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"using QuasinormalModes\nusing SymEngine","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","title":"Complete Example: The Harmonic Oscillator","text":"Next, we create a parametric type that sub-types AnalyticAIMProblem. As the eigenvalue in the master equation is a quadratic polynomial, we will sub-type QuadraticEigenvalueProblem with the following structure:","category":"page"},{"location":"sho/","page":"Complete Example: The Harmonic Oscillator","t
[…]
