+++ src/eigenvalue_computing/compute_eigenvalues.jl
@@ -23,22 +23,22 @@ function computeEigenvalues(
     m::AIMSteppingMethod,
     p::QuadraticEigenvalueProblem{N,T},
     c::AIMCache{N,Polynomial{T}};
-    plr_polish::Bool=true, 
-    plr_epsilon::Real=1.0e-10
-    ) where {N <: Unsigned,T <: Number}
-    
+    plr_polish::Bool = true,
+    plr_epsilon::Real = 1.0e-10
+) where {N<:Unsigned,T<:Number}
+
     # Compute the AIM "quantization condition"
     δ = computeDelta!(m, p, c)
 
     # Solve the quantization condition to obtain the eigenvalues
-    eigenvalues = PolynomialRoots.roots(coeffs(δ), polish=plr_polish, epsilon=plr_epsilon)
-    
+    eigenvalues = PolynomialRoots.roots(coeffs(δ), polish = plr_polish, epsilon = plr_epsilon)
+
     if isempty(eigenvalues)
-        println("Warning: The computed mode array is empty. This means that no roots of the polynomial equation in ω were found.")
+        @warn "The computed mode array is empty. This means that no roots of the polynomial equation in ω were found."
     end
 
     return eigenvalues
-end 
+end
 
 """
     computeEigenvalues(
@@ -69,10 +69,10 @@ function computeEigenvalues(
     p::NumericAIMProblem{N,T},
     c::AIMCache{N,T},
     guess::T;
-    nls_xtol::Real=1.0e-10,
-    nls_ftol::Real=1.0e-10,
-    nls_iterations::Int=1000
-    ) where {N <: Unsigned,T <: Complex}
+    nls_xtol::Real = 1.0e-10,
+    nls_ftol::Real = 1.0e-10,
+    nls_iterations::Int = 1000
+) where {N<:Unsigned,T<:Complex}
 
     # This function is passed to NLsolve to find the roots of δ
     function f!(F, x)
@@ -80,9 +80,9 @@ function computeEigenvalues(
         F[1] = real(y)
         F[2] = imag(y)
     end
-    
+
     # compute the roots of δ using NLsolve
-    return nlsolve(f!, [real(guess), imag(guess)], ftol=nls_ftol, xtol=nls_xtol, iterations=nls_iterations)
+    return nlsolve(f!, [real(guess), imag(guess)], ftol = nls_ftol, xtol = nls_xtol, iterations = nls_iterations)
 end
 
 """
@@ -120,23 +120,23 @@ function computeEigenvalues(
     p::NumericAIMProblem{N,T},
     c::AIMCache{N,T},
     guess::T;
-    roots_atol::Real=1.0e-10,
-    roots_rtol::Real=1.0e-10,
-    roots_xatol::Real=1.0e-10,
-    roots_xrtol::Real=1.0e-10,
-    roots_maxevals::Int=100
-    ) where {N <: Unsigned,T <: Real}
+    roots_atol::Real = 1.0e-10,
+    roots_rtol::Real = 1.0e-10,
+    roots_xatol::Real = 1.0e-10,
+    roots_xrtol::Real = 1.0e-10,
+    roots_maxevals::Int = 100
+) where {N<:Unsigned,T<:Real}
 
     try
         find_zero(
             x -> computeDelta!(m, p, c, x),
             guess,
-            atol=roots_atol,
-            rtol=roots_rtol,
-            xatol=roots_xatol,
-            xrtol=roots_xrtol,
-            maxevals=roots_maxevals
-            )
+            atol = roots_atol,
+            rtol = roots_rtol,
+            xatol = roots_xatol,
+            xrtol = roots_xrtol,
+            maxevals = roots_maxevals
+        )
     catch
         println("find_zeros was unable to converge to an eigenvalue")
     end
