### 2022-02-08 #4077 @lucass-carneiro
Hi @cescalara,

Thank you for your review and comments. They were all very helpful and have improved the quality of the package. Here is a brief list of changes made in the "paper" and "gh-pages" branches to address your comments:

1. Both the documentation and the tests have now their own `Project.toml` and `Manifest.toml` files with their dependencies tracked.
2. I have found no other place in the code that outputs logging-like messages with `println` throughout the code and replaced the spot you indicated with `@warn`. I still use `println` in places where I want to format the output a certain way and `error` where appropriate.
3. The documentation is now synced with the examples. I have also made the package loading step of the examples a little bit more explicit in the documentations. `Literate.jl` seems a great tool, but (at least for now) I thought it better to keep tracking things manually in order to address other issues more quickly, since integrating `Literate.jl`'s workflow with the existing structure would require some effort.
4. The benchmark data, as well as the source code necessary to generate it and the plots that are included in the paper are now in the repository [here](https://github.com/lucass-carneiro/QuasinormalModes.jl/tree/paper/benchmark). I have included a small `README.md` file with instructions on how to generate new benchmarks.
5. I have added a new section in the documentation with general tips, workflow guidelines and how to choose parameters in the AIM [here](https://lucass-carneiro.github.io/QuasinormalModes.jl/org/#Computing-eigenvalues-and-general-workflow-guidelines).
6. About the citations, are you referring to the act that all citations have parenthesis around them? If so, I thought that this style was accepted in JOSS since i have obtained it from [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
7. "Oscillator" is now spelt correctly throughout the code and source file names.
8. I will consider changing the names from camel to snake case, however since this is a major API change It will probably take some time to change everything and I'm not sure that the review is a good period for that type of overhaul.
