### 2022-02-11 #4077 @pdebuyl
Thanks @cescalara for the review so far!

@lucass-carneiro JOSS appreciates to see the best practices for a specific field or language but the choice of camel case vs snake case is not relevant for the review :-)
