### 2022-02-28 #4077 @cescalara
Hi @lucass-carneiro,

I have been through your response and changes to the code, thanks a lot for taking my comments on board. I confirm that I was able to run your benchmark example and produce equivalent plots to those shown in the paper, so I haved checked off the box regarding performance claims. I also think the added documentation is very helpful. I add a few more minor comments/replies below.

* When using `computeEigenvalues` in the Schwarzschild numerical example, I get an `NLSolve.SolverResults` output as explained in the docs. Most fields in the output are self-explanatory, but it might be nice to have a example of what the output values could be for users to compare to. I notice that the Inf-norm is a large number and not sure if this is cause for concern in this case (although I admit this may be more obvious to a reviewer with expertise in these methods).

* You might consider using [deploydocs](https://juliadocs.github.io/Documenter.jl/v0.9/lib/public.html#Documenter.deploydocs) so that you docs track different versions/releases of the software.

> About the citations, are you referring to the act that all citations have parenthesis around them? If so, I thought that this style was accepted in JOSS since i have obtained it from here.

This is a really minor point, but I meant that for citations that are part of the sentence (e.g. line 40 in your paper) one usually drops the parentheses.
