### 2022-03-09 7d4c240
Added automatic doc deployment with GitHub actions
