### 2022-03-09 7d988ae
Reset deoc deployment settings to default values
