### 2022-03-09 7e3c1a2
Cleaned up documentation folder for new automatic doc deployment
