### 2022-03-09 #4077 @cescalara
Hi @lucass-carneiro,

Thanks again for taking all the comments on board.

> There are links in REDME.md in master that point to the stable and development versions of the docs, although the stable link is not working. I suspect that when I tag the next release version, Documenter.jl will auto generate the documentation for the new version and create the stable link. I suspect it did not do so now because I have triggered it by simple pushes to master that contained only preexisting tags.

That sounds right, although it should be able to pick up existing tags. If there are still any issues after tagging your new release, you can always just clear out the `gh-pages` branch and rerun the action (at least this has worked for me).

> I have removed parenthesis around the citations that are part of a sentence. I tried to get whedon to compile a new preview but it did not seem to work, I don't know why. Maybe @pdebuyl needs to issue the command? In any case, it is minor change but I think it is best to update the paper preview.

No worries, this really isn't a big deal.

> do you think that including the SolverResults output in the documentations for the examples and briefly going over what the parameters are would help?

That would be great, I think then I can check all the boxes!
