### 2022-03-09 #4077 @pdebuyl
Hi @lucass-carneiro I did not consider that @JamieBamber might not read his notifications from GitHub, which happens not so unfrequently with JOSS reviewers. It was the case and he confirmed today that he would review. Sorry for the delay about this!
