### 2022-03-09 #4077 @pdebuyl
Regarding the citations, there is a mismatched curly brace at https://github.com/lucass-carneiro/QuasinormalModes.jl/blob/paper/paper/paper.bib#L131 so the bibtex parsing fails.
