### 2022-03-11 dddc461
Added logo to readme.md
