### 2022-03-11 fdf6ee1
Added new examples teaching how to use RootsAndPoles.jl to finde eigenvalues. Added a readme file in the examples folder to aind navigation
