+++ examples/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+RootsAndPoles = "a0859a10-ccb4-11e8-145d-674b3750773a"
+SymEngine = "123dc426-2d89-5057-bbad-38513e3affd8"
