### 2022-03-11 93373e9
Added link to examples folder in readme file.
