### 2022-03-11 629c983
Added missing . in examples readme file
