### 2022-03-11 f334a15
Added link to examples using RootsAndPoles.jl
