### 2022-03-12 7cd4ad9
Added new documentation bits explaning SolverResults results, advice on how to proceed when the inf norm is large, how to find initial guesses etc
