### 2022-03-12 #4077 @lucass-carneiro
Hi @cescalara

I have made a few additions both of code examples and documentation to the package. Since I think all of these changes are important to the package as a whole, the have been pushed directly to `master`.

1. There are two new examples, [here](https://github.com/lucass-carneiro/QuasinormalModes.jl/blob/master/examples/harmonic_oscillator_roots_and_poles.jl) and [here](https://github.com/lucass-carneiro/QuasinormalModes.jl/blob/master/examples/schwarzschild_roots_and_poles.jl). These examples show how the user is not bound to use `NLsolve.jl` and is in fact free to choose any root finding method they wish. In this particular case, I'm using [`RootsAndPoles.jl`](https://github.com/fgasdia/RootsAndPoles.jl). These new examples are referenced multiple times across the documentation, when appropriate. There is also a new `README.md` file in the examples folder and a `Projec.toml` and `Manifest.toml` file as well in order to keep track of example dependencies separately and clearly.
2. There is a new note marked in blue at the bottom of the page [here](https://lucass-carneiro.github.io/QuasinormalModes.jl/dev/org/#Computing-eigenvalues-and-general-workflow-guidelines) giving advice on how to find initial guesses for physics problem where one might have no idea where to start. I have decided to add this advice even tough you didn't ask for it directly as I realized that this particular topic was missing from the rest of the documentation.
3. There is a new section [here](https://lucass-carneiro.github.io/QuasinormalModes.jl/dev/schw/#Interpreting-SolverResults-output) that deals with interpreting `SolverResult` objects and a brief discussion about the large inf-norm. To summarize, my advice in these cases is to always double check with a different root finding method. In the Schwarzschild case, `RootsAndPoles.jl` is also able to find the same root by a completely different method that does not rely on polishing. which indicates that the value is in fact a root of the quantization condition and a quasinormal mode candidate. Note that in the harmonic oscillator example, `NLsolve.jl` converges by both criteria, probably because this example is substantially simpler than the Schwarzschild quasinormal modes.

Regarding changes made to the `paper` branch:

1. @pdebuyl I have fixed the paper compilation issues and tested the compilation locally using the docker image. For good measure, I have decided to also include the generated PDF. About @JamieBamber's review, don't worry about the delay. It is good that you managed to get a hold of him and that he is able to do the review. I have no trouble waiting.
2. I have added a few people to the `Acknowledgements` section of the paper.
