### 2022-03-13 #4077 @cescalara
Hi @lucass-carneiro,

Thank you for the further additions, I think these together help to show the subtleties of using your package in a robust way. I confirm the new examples run without issue and that I can reproduce the results shown in the documentation.

From my side I think I can conclude the review and I look forward to the thoughts of @JamieBamber.
