### 2022-03-16 #4077 @JamieBamber
@lucass-carneiro
Apologies for the delay in my review. The work looks very impressive. I agree with the comments of @cescalara, and I only had a few minor additional comments:

1.  You discuss the need to carefully refine the results to distinguish true eigenvalues from spurious ones, and to find others not discovered from that x_0. In your research work have you developed a systematic method or code to do this, and if so could it be included in the repository?

2.  In the Schwarzschild BH example after seeing the output list of candidate frequencies found from the analytic problem I immediately wanted to plot a graph to see what they look like and compare with the known plots of Schwarzschild QNM frequencies. Hence, I was wondering if it might be a good idea to add such a plotting feature to that example, so that users can see the results visually and not merely from a list of numbers, as well as a comparison plot of the true QNM frequencies to visually demonstrate the method works?

3. There are a few minor typos in the documentation, such as using "Weather" instead of "Whether" at the bottom of the "Complete Example: Schwarzschild Quasinormal Modes" page.

4. Why did you chose the AIM method for your code over the alternatives?

5. How does the time scaling and error convergence of this code compare to the existing alternatives (where comparable) or non-AIM methods?
