### 2022-04-06 db31ec2
Added new example for computing Schwazschild QNMs up to large values of l
