+++ examples/high_l_Schwarzschild/guesses.jl
@@ -1,0 +1,900 @@
+# ------------------------------------------------------------------
+# Reference QNMs for use as initial guesses
+#
+# Format: (l, n, real part, imaginary part)
+# ------------------------------------------------------------------
+
+const spin_0 = [
+    (0, "0", "0.1104", "-0.1048"),
+    (1, "0", "0.2911", "-0.0980"),
+    (2, "0", "0.4832", "-0.0968"),
+    (2, "1", "0.4632", "-0.2958"),
+    (3, "0", "0.6752", "-0.0965"),
+    (3, "1", "0.6604", "-0.2923"),
+    (3, "2", "0.6348", "-0.4941"),
+    (4, "0", "0.8673", "-0.0964"),
+    (4, "1", "0.8857", "-0.2909"),
+    (4, "2", "0.8345", "-0.4895"),
+    (4, "3", "0.8064", "-0.6926"),
+    (5, "0", "1.0585", "-0.096225"),
+    (5, "1", "1.0585", "-0.28868"),
+    (5, "2", "1.0585", "-0.48113"),
+    (5, "3", "1.0585", "-0.67358"),
+    (5, "4", "1.0585", "-0.86603"),
+    (5, "5", "1.0585", "-1.0585"),
+    (6, "0", "1.2509", "-0.096225"),
+    (6, "1", "1.2509", "-0.28868"),
+    (6, "2", "1.2509", "-0.48113"),
+    (6, "3", "1.2509", "-0.67358"),
+    (6, "4", "1.2509", "-0.86603"),
+    (6, "5", "1.2509", "-1.0585"),
+    (6, "6", "1.2509", "-1.2509"),
+    (7, "0", "1.4434", "-0.096225"),
+    (7, "1", "1.4434", "-0.28868"),
+    (7, "2", "1.4434", "-0.48113"),
+    (7, "3", "1.4434", "-0.67358"),
+    (7, "4", "1.4434", "-0.86603"),
+    (7, "5", "1.4434", "-1.0585"),
+    (7, "6", "1.4434", "-1.2509"),
+    (7, "7", "1.4434", "-1.4434"),
+    (8, "0", "1.6358", "-0.096225"),
+    (8, "1", "1.6358", "-0.28868"),
+    (8, "2", "1.6358", "-0.48113"),
+    (8, "3", "1.6358", "-0.67358"),
+    (8, "4", "1.6358", "-0.86603"),
+    (8, "5", "1.6358", "-1.0585"),
+    (8, "6", "1.6358", "-1.2509"),
+    (8, "7", "1.6358", "-1.4434"),
+    (8, "8", "1.6358", "-1.6358"),
+    (9, "0", "1.8283", "-0.096225"),
+    (9, "1", "1.8283", "-0.28868"),
+    (9, "2", "1.8283", "-0.48113"),
+    (9, "3", "1.8283", "-0.67358"),
+    (9, "4", "1.8283", "-0.86603"),
+    (9, "5", "1.8283", "-1.0585"),
+    (9, "6", "1.8283", "-1.2509"),
+    (9, "7", "1.8283", "-1.4434"),
+    (9, "8", "1.8283", "-1.6358"),
+    (9, "9", "1.8283", "-1.8283"),
+    (10, "0", "2.0207", "-0.096225"),
+    (10, "1", "2.0207", "-0.28868"),
+    (10, "2", "2.0207", "-0.48113"),
+    (10, "3", "2.0207", "-0.67358"),
+    (10, "4", "2.0207", "-0.86603"),
+    (10, "5", "2.0207", "-1.0585"),
[…]
