+++ examples/Project.toml
@@ -1,3 +1,8 @@
 [deps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 RootsAndPoles = "a0859a10-ccb4-11e8-145d-674b3750773a"
 SymEngine = "123dc426-2d89-5057-bbad-38513e3affd8"
