### 2022-04-06 dcadd01
Added a documentation section about the new examples and plots
