### 2022-04-06 c93412a
Fixed documentation typos, regenerated example plot
