### 2022-04-06 aa595fa
added code for showing plot in docs
