### 2022-04-06 e95863c
Changed the plot script to produce svg files directly. Updated the plot
