### 2022-04-06 e05dbf6
Fixed \ell not being shown in plot legend
