### 2022-04-06 #4077 @lucass-carneiro
Hello @JamieBamber. Once again, I apologize my long delay in answering your questions. As I have said, I was swamped with other duties and did not find the time to answer you properly.

In regards to your comments:

1. So far I have not coded any automatic solution for finding `x_0`. As I've mentioned in the docs, I usually start halfway through in the compactifyied domain (to me this usually means `x_0 = 0.5`) and start perturbing this parameter from there. In practice, I never found the need to deviate much further from this value, as `x_0` speeds up (or slows down) the convergence of the mode and does not change the set of eigenvalues obtained, at least from a theoretical standpoint. In order to develop a systematic way of fine tuning `x_0`, the user could take, for instance, the time required to find a certain mode and choose `x_0` such as to minimize this time. This is not a difficult thing to do, but again, for most users and practical purposes unnecessary for good usage of the method.
2. It is very easy to plot the obtained eigenvalues, as they are simple Julia arrays. I have not included such functionality because plot aesthetics are a highly subjective topic so I have chosen to give users data instead of ready made plots. Nevertheless, I also think that a plot of the modes would be of interest. That is why I have added a new section in the documentation that can be found [here](https://lucass-carneiro.github.io/QuasinormalModes.jl/dev/schw/#Computing-high-\\ell-quasinormal-modes). The code used in this example is an excerpt of a code that I'm currently using (together with collaborators) in a paper that compares the AIM with pseudo-spectral methods for several bosonic and fermionic Schwarzschild perturbations. We have found so far that even for large ``\ell`` values the results computed with both methods agree to machine precision (when using `BigFloats` for the computations) and also agree with available literature values. We are also in the processes of discovering and cross-comparing several modes not listed in existing references. Initial guesses in these computations were produced using the analytic formula that yields the Schwarzschild QNMs for large ``\ell`` and the list of modes can be found [here](https://github.com/lucass-carneiro/QuasinormalModes.jl/blob/master/examples/high_l_Schwarzschild/qnm_2022-04-05%2017:49:10_s_0.dat)
3. Noted. I have combed the file and fixed a few other typos.
4. In my opinion, the main advantage of the AIM over other methods is it's flexibility and generality. It's formulated in such a way that it abstracts away the specific features of problems and presents a general solution to the second order ODE of interest. I have written the package in a way that attempts to expose that abstraction to the user: Give me your ODE coefficients and I shall give you eigenvalues. It does not matter if you are trying to find the modes of an extreme BH or if your eigenvalues appears inside a complicated function in one of the ODE's coefficients. The method works regardless of such details. In contrast, let us look at another popular method of solving the same problem: Leaver's continued fraction method. This method is based upon finding a continued fraction expression from the ODE of interest and searching for roots of said fraction to determine the problem's eigenvalues. The difficulty lies in the construction of the continued fraction: To do so, one must derive a recurrence relation for series expansion coefficients which may have a varying number of terms. Each problem, generates a uniquely different recurrence relation and it's hard to generalize this and abstract away certain problem details. Furthermore, the method needs to be supplemented with additional techniques in the case of an extreme black hole (see [this paper](https://arxiv.org/abs/1409.7440)). These requirements make it unsuitable for a general solver. Another popular candidates are pseudo-spectral methods, which discretize the derivative operators and transforms the problem into a matrix-generalized eigenvalue problem. This transformation is straightforward if the eigenvalue being sought is a liner or polynomial function. It is not possible (at least to my knowledge) to pose a generalized eigenvalue problem using a spectral method if the eigenvalue is any other type of function. If for instance the eigenvalue being sought appears inside a spheroidal harmonic, pseudo-spectral methods are not an option. The AIM does not care in general for what your ODE coefficients look like. You simple need to be able to evaluate them inside your compactifyied domain. As I intended to have a general eigenvalue solver, the AIM was a natural candidate.
5. That is a question that is difficult to answer. The AIM is a relatively new method (circa 2003) and only very recently (see [this paper](https://arxiv.org/pdf/2003.06730.pdf)) things like success and failure conditions and why the method actually works robustly were understood formally. As the field progresses I intend to keep up with such developments and keep the underlying math in the package up to date. That being said, I prefer not to make any specific claims about the method's convergence rates in comparison to other methods as they will probably be wrong. It can be shown empirically, as I have, that the method does converge but that in some cases it oscillates and never converges (see the paper I've indicated). Nevertheless, the conditions for success and failure are now understood and the robustness of the method can be mathematically proven. I might one day attempt to make such comparisons but that would be the subject of an entirely new work.
