### 2022-04-20 #4077 @pdebuyl
Thanks @JamieBamber for confirming your acceptance!

@lucass-carneiro I'll proceed soon to the next editorial steps https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
