### 2022-04-22 fb78ba3
Added the JOSS paper to the main branch
