+++ paper/paper.bib
@@ -1,0 +1,136 @@
+@article{qnmspectral,
+    author = "Jansen, Aron",
+    title = "{Overdamped modes in Schwarzschild-de Sitter and a Mathematica
+             package for the numerical computation of quasinormal modes}",
+    journal = "Eur. Phys. J. Plus",
+    volume = "132",
+    year = "2017",
+    number = "12",
+    pages = "546",
+    doi = "10.1140/epjp/i2017-11825-9",
+    eprint = "1709.09178",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc"
+}
+
+@article{spectralbp,
+    author = "Fortuna, Sean and Vega, Ian",
+    title = "{Bernstein spectral method for quasinormal modes and other
+             eigenvalue problems}",
+    eprint = "2003.06232",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    month = "3",
+    year = "2020",
+}
+
+@misc{bhpt_quasinormalmodes,
+    author = {O'Toole, Connor and Macedo, Rodrigo and Stratton, Tom and Wardell,
+              Barry},
+    title = {QuasiNormalModes},
+    year = {2019},
+    publisher = {GitHub},
+    journal = {GitHub repository},
+    url = {https://github.com/BlackHolePerturbationToolkit/QuasiNormalModes},
+}
+
+@article{bhpt_qnm,
+    author = "Stein, Leo C.",
+    title = "{qnm: A Python package for calculating Kerr quasinormal modes,
+             separation constants, and spherical-spheroidal mixing coefficients}",
+    journal = "J. Open Source Softw.",
+    volume = "4",
+    year = "2019",
+    number = "42",
+    pages = "1683",
+    doi = "10.21105/joss.01683",
+    eprint = "1908.10377",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc"
+}
+
+@article{aim_original,
+    doi = {10.1088/0305-4470/36/47/008},
+    url = {https://doi.org/10.1088/0305-4470/36/47/008},
+    year = 2003,
+    month = {nov},
+    publisher = {{IOP} Publishing},
+    volume = {36},
+    number = {47},
+    pages = {11807--11816},
+    author = {Hakan Ciftci and Richard L Hall and Nasser Saad},
+    title = {Asymptotic iteration method for eigenvalue problems},
+    journal = {Journal of Physics A: Mathematical and General},
+    abstract = {An asymptotic iteration method for solving second-order
+                homogeneous linear differential equations of the form y″ = λ0(x)
+                y′ + s0(x)y is introduced, where λ0(x) ≠ 0 and s0(x) are C∞
+                functions. Applications to Schrödinger-type problems, including
+                some with highly singular potentials, are presented.},
+}
+
+@article{aim_improved,
+    author = "Cho, H. T. and Cornell, A. S. and Doukas, Jason and Huang, T. R.
+              and Naylor, Wade",
+    title = "{A New Approach to Black Hole Quasinormal Modes: A Review of the
+             Asymptotic Iteration Method}",
+    eprint = "1111.5024",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    reportNumber = "YITP-11-97, WITS-CTP-83, OU-HET-735-2011",
+    doi = "10.1155/2012/281705",
+    journal = "Adv. Math. Phys.",
+    volume = "2012",
+    pages = "281705",
+    year = "2012",
+}
+
+@misc{berti_bangalore,
+    author = {Berti, Emanuele},
[…]
