+++ paper/paper.md
@@ -1,0 +1,76 @@
+---
+title: '`QuasinormalModes.jl`: A Julia package for computing discrete eigenvalues of second order ODEs'
+tags:
+  - Julia
+  - Differential equations
+  - Black holes
+  - Discrete eigenvalues
+authors:
+  - name: Lucas Timotheo Sanches
+    orcid: 0000-0001-6764-1812
+    affiliation: 1 
+affiliations:
+ - name: Centro de Ciências Naturais e Humanas, Universidade Federal do ABC (UFABC)
+   index: 1
+date: 17 September 2021
+bibliography: paper.bib
+
+---
+
+# Summary
+
+In General Relativity, when perturbing a black hole with an external field, or particle, the system relaxes by emitting damped gravitational waves, known as *quasinormal modes*. These are the characteristic frequencies of the black hole and gain the *quasi* prefix due to the fact that they have a real frequency, which represents the oscillation frequency of the response, and an imaginary frequency that represents the decay rate of said oscillations. In many cases, such perturbations can be described by a second order homogeneous ordinary differential equation (ODE) with discrete complex eigenvalues.
+
+Determining these characteristic frequencies quickly and accurately for a large range of models is important for many practical reasons. It has been shown that the gravitational wave signal emitted at the final stage of the coalescence of two compact objects is well described by quasinormal modes [@buonanno; @seidel]. This means that if one has access to a database of quasinormal modes and of gravitational wave signals from astrophysical collision events, it is possible to characterize the remnant object using its quasinormal frequencies. Since there are many different models that aim to describe remnants, being able to compute the quasinormal frequencies for such models in a reliable way is paramount for confirming or discarding them.
+
+# Statement of need
+
+`QuasinormalModes.jl` is a `Julia` package for computing the quasinormal modes of any General Relativity model whose perturbation equation can be expressed as second order  homogeneous ODE. Not only that, the package can be used to compute the discrete eigenvalues of *any* second order homogeneous ODE (such as the energy eigenstates of the time independent Schrödinger equation) provided that these eigenvalues actually exist. The package features a flexible and user friendly API where the user simply needs to provide the coefficients of the problem ODE after incorporating boundary and asymptotic conditions on it. The user can also choose to use machine or arbitrary precision arithmetic for the underlying floating point operations involved and whether or not to do computations sequentially or in parallel using threads. The API also tries not to force any particular workflow on the users so that they can incorporate and adapt the existing functionality on their research pipelines without unwanted intrusions. Often user friendliness, flexibility and performance are treated as mutually exclusive, particularly in scientific applications. By using `Julia` as an implementation language, the package can have all of theses features simultaneously.
+
[…]
