+++ Project.toml
@@ -1,7 +1,7 @@
 name = "QuasinormalModes"
 uuid = "7b4b3212-f72a-4ba2-afa8-103f0bfb7cd6"
 authors = ["Lucas Timotheo Sanches <lucas.t@ufabc.edu.br>"]
-version = "1.0.0"
+version = "1.1.0"
 
 [deps]
 NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
