+++ README.md
@@ -11,6 +11,8 @@ To compute eigenvalues (and thus quasinormal frequencies) this package uses the 
 ![Build Status](https://github.com/lucass-carneiro/QuasinormalModes.jl/actions/workflows/CI.yml/badge.svg)
 [![codecov](https://codecov.io/gh/lucass-carneiro/QuasinormalModes.jl/branch/master/graph/badge.svg?token=GK9052NQK2)](https://codecov.io/gh/lucass-carneiro/QuasinormalModes.jl)
 
+[![DOI](https://zenodo.org/badge/316366566.svg)](https://zenodo.org/badge/latestdoi/316366566)
+
 # Author
 [Lucas T. Sanches](lucas.t@ufabc.edu.br), Centro de Ciências Naturais e Humanas, Universidade Federal do ABC (UFABC).
 
