+++ v1.1.0-Zenodo/siteinfo.js
@@ -1,0 +1,1 @@
+var DOCUMENTER_CURRENT_VERSION = "v1.1.0-Zenodo";
