### 2022-04-22 #4077 @lucass-carneiro
@pdebuyl As per the instructions, I have tagged a new release (v1.1.0) and archived it in Zenodo. Here is the [DOI](https://doi.org/10.5281/zenodo.6478503).

I have also move the paper from the `paper` branch to the `master` branch and deleted the `paper` branch.
