### 2022-05-19 #4077 @pdebuyl
@lucass-carneiro can you edit the metadata of the zenodo archive so that it matches the title and author list of the article?

Also, can you spell out "Journal of Open Source Software" in the bibliography for Stein 2019? Apart for the "J" of "Journal", I don't think that the words can be shortened.

I only ran into one problem while testing the software, it is that I had to explicitly install SymEngine after having installed QuasiNormalModes. It would make sense to make it a run-time dependency but I don't know enough about Julia packaging to know if this is normal or not.
