### 2022-05-19 7efaf86
Updated citation to display full journal name
