+++ paper/paper.bib
@@ -38,7 +38,7 @@
     author = "Stein, Leo C.",
     title = "{qnm: A Python package for calculating Kerr quasinormal modes,
              separation constants, and spherical-spheroidal mixing coefficients}",
-    journal = "J. Open Source Softw.",
+    journal = "Journal of Open Source Software",
     volume = "4",
     year = "2019",
     number = "42",
