### 2022-05-19 #4077 @lucass-carneiro
@pdebuyl
* I have edited the Zenodo archive metadata to match the paper information.
* JOSS is now spelt without abbreviations in the Stein 2019 citation.
* If you are trying to run the tests or examples, activate the project environment using the `--project` flag. Since the tests and examples track their dependencies separately from the main package, everything should work fine without the need to manually download `SymEngine`. If the user is running outside of the `project` environment (which is often the case) and trying to define an analytic problem then `SymEngine` will need to be manually added by the user, even tough it is explicitly defined as a dependency of the main package. There are ways to automate this. If the user is using a `Pluto` notebook for instance, `using` statements will trigger an automatically installation if the packages are not found in the local registry.
