### 2022-05-20 c2006dd
Updated the paper with figures and added thanks to contributors
