### 2022-05-20 #4077 @lucass-carneiro
@pdebuyl FYI, I have updated the paper to correct an issues where benchmark figures were missing and added new thanks to contributors
