+++ paper/paper.bib
@@ -84,18 +84,6 @@
     year = "2012",
 }
 
-@misc{berti_bangalore,
-    author = {Berti, Emanuele},
-    title = {Black Hole Perturbation Theory},
-    year = {2016},
-    publisher = {Dropbox},
-    journal = {Lecture notes},
-    url = {
-           https://www.dropbox.com/sh/9th1um175m8gco9/AACCIkvNa3h-zdHBMZkQ31Baa?dl
-           =0&preview=Bangalore.pdf},
-}
-
-
 @article{berti_ringdown,
     author = "Berti, Emanuele and Cardoso, Vitor and Starinets, Andrei O.",
     title = "{Quasinormal modes of black holes and black branes}",
