### 2022-05-23 #4077 @lucass-carneiro
@pdebuyl I have removed the extraneous citation. Thanks for pointing it out.
